+++ src/utils.ts
@@ -25,6 +25,10 @@ import {
   ESDTTransferPayloadBuilder,
   ESDTNFTTransferPayloadBuilder,
 } from '@multiversx/sdk-core';
+import decompress from 'decompress';
+import getStream from 'get-stream';
+import got, { ResponseType } from 'got';
+import { pEvent } from 'p-event';
 import axios, { AxiosResponse } from 'axios';
 import { parseUserKey, UserSigner } from '@multiversx/sdk-wallet';
 import {
@@ -41,6 +45,7 @@ import {
   writeFileSync,
   promises,
 } from 'fs';
+import path from 'path';
 import { exit, cwd } from 'process';
 import {
   apiProvider,
@@ -961,4 +966,46 @@ export const distributeSftSingleAddress = async (
       txStatus: 'failed',
     };
   }
+};
+
+// Based on not maintained: https://github.com/kevva/download (simplified)
+export const downloadAndExtract = (
+  uri: string,
+  output: string | null,
+  options: Record<string, unknown>
+) => {
+  const gotOptions = {
+    responseType: 'buffer' as unknown as ResponseType,
+    https: {
+      rejectUnauthorized: process.env.npm_config_strict_ssl !== 'false',
+    },
+  };
+
+  const stream = got.stream(uri, gotOptions);
+
+  const promise = pEvent(stream, 'response')
+    .then((res) => {
+      return Promise.all([
+        getStream(stream, { encoding: 'buffer' as BufferEncoding }),
+        res,
+      ]);
+    })
+    .then((result) => {
+      const [data] = result;
+
+      if (!output) {
+        return decompress(data, options);
+      }
+
+      const filename = options.filename;
+      const outputFilepath = path.join(output, filename as string);
+
+      return decompress(data, path.dirname(outputFilepath), options);
+    });
+
+  return {
+    then: promise.then.bind(promise),
+    catch: promise.catch.bind(promise),
+    ...stream,
+  };
 };